Fix 64-bit SmallInteger encoding/decoding with tag bits in low bits
- Tag bits are in LOW bits (bottom 2 bits) - standard tagged pointer approach
- Fixed SmallInteger decoding to use arithmetic right shift for proper sign extension
- Changed SmallInteger interface from int32_t to int64_t (62-bit value range)
- Updated test to use correct 62-bit signed range (2^61 - 1 to -2^61)
- All TaggedValue tests now pass (6/6)
- Tag bits: bits 0-1, Value: bits 2-63 (62 bits)

+++ tests/bytecode/bytecode_test_helpers.cpp
@@ -121,7 +121,7 @@ size_t getStackDepth(Context* context) {
     return context->stack.size();
 }
 
-TaggedValue makeSmallInteger(int32_t value) {
+TaggedValue makeSmallInteger(int64_t value) {
     return TaggedValue::fromSmallInteger(value);
 }
